lab/lab2/analyzer/requirements.txt
etw-parser
pillow